modify timer ui
