+++ assets/ui/Timer.ui
@@ -17,7 +17,7 @@
                     "use-content-height" : true,
                     "position-horizontal-center" : {},
                     "position-top" : {
-                        "offset": 70
+                        "offset": 80
                     }
                 }
             }
