+++ src/main/java/org/terasology/ligthandshadow/componentsystem/events/GameStartMessageEvent.java
@@ -1,0 +1,10 @@
+// Copyright 2021 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
+package org.terasology.ligthandshadow.componentsystem.events;
+
+import org.terasology.engine.entitySystem.event.Event;
+import org.terasology.engine.network.OwnerEvent;
+
+@OwnerEvent
+public class GameStartMessageEvent implements Event {
+}
