+++ src/main/java/org/terasology/ligthandshadow/componentsystem/events/RemoveInvulnerabilityEvent.java
@@ -1,0 +1,9 @@
+// Copyright 2021 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
+package org.terasology.ligthandshadow.componentsystem.events;
+
+import org.terasology.engine.entitySystem.event.Event;
+
+public class RemoveInvulnerabilityEvent implements Event {
+
+}
